Added Post selection and Reply button
Switched to DataView/DataItem structure for the forum posts
structure. The new view ForumPost is the DataItem structure
containing the user post. The post structure is contains
the user avatar, the firstname, the lastname, the subject,
the message and a reply button. The Reply button is hidden by
default. It is shown when a post is selected. Selecting another
post hides the button in the previously selected one.
All posts are indented just like in the previous implementation.
The css classes have been renamed though. The new convention used
is going to be class-selector-name and id_selector_name.