Added User controller
Added User controller. At the moment the only function
of this controller is addNewUser. This function queries
the data of the new user and stores it locally.

The code that loads and pre-processes the post list code
has been re-factored. The new code checks if there are
new users before formating the list. It is possible to
have new posts belonging to users that still haven't been
added to the users store. Getting the new users' data first
avoids having empty user data fields in the new posts.