BUGFIX: Updated modules notification was broken
The update modules notification was broken. The updates were
silent because of how the sencha framework manages instances.
The fundamental problem is that two different instances
of the same Model with the same id are just one object in the
cache. So the first instance is overwritten by the second.
So everytime new updates were asked the localstorage was
overwritten by the updates automatically. This way was impossible
to understand which modules were updated and which not.

In this fix the id for every instance is generated by using the
idproperty: uuid. The id is defined as localid.

Only Course and User model are still using id field and not localid.