Upload Assignment Draft Support
The upload of multiple files is ready. Choosing the files is not
ready yet because of a sencha bug affecting the xtype:file
component.

So this is is only a draft. Some checks are still missing
such as the maximum number of files to upload in the assignment.
These checks are going to be implemented after the bug affecting
the xtype:file component is fixed.