Bugfixes:
Module entries were cloned due to the filterBy function and the
asynchronous updating of the moodle store.

Minor fix for the Choice activity.

Fixed the DataItem null records issue. This happened when pressing
the back button, the updateRecord was called but the parameter
was null. It was probably due to the filtering of the stores.