Integrated the Scorm Reader
Raw integration of the Scorm reader developed by
Nicola Rizzo (http://www.nicolarizzo.com).

The current version points to an internal demo scorm which
for the copyright issues is not pubblished in this commit.
There are still some problems with the ExtractZipFile module.
Not all kind of scorm archives are extracted correctly so that is
why the Scorm reader reads the internal, demo, content.