Bugfix for Resource modules containing html files
This commit fixes the problem with Resources containing html files.
The previous mechanism used to download files and then open them.
This was a problem with html files which can contain links to other
files on the platform. The current mechanism opens a InAppBrowser
window and opens the mod/resource/view.php page on the platform.
This way the internal links work as supposed.