Extract course section labels info from the server
This commit bring the support for the course sections.
Now all the course sections info is extracted from the server.
The sections can have the mixed title now depending on the
section label settings on the platform.

The new order in the ModulesList reflects the one on the platform.
The modules are not grouped by the type anymore.