New file selecting mechanism for Android platforms
Implemented a new file selecting mechanism for Android
platforms. Android 4.4 removed the support for the input tag
type file in the Chrome WebView object.
This commit switches to the fileChooser cordova plugin.

A new File object has been added. This one is completely new
and does not use the sencha's filefield xtype. The new object
is named FileSlot. It consists only of the view that is used
for loading the picked file. The file picking is actually done
by the new FileSystem singleton.

All other controllers use the new FileSystem singleton except
the supsi code. This has been left as it is for legacy purposes.
Even the supsi code is going to be changed sooner or later though.

A new cordova plugin has been added:
org.apache.cordova.file-system-roots

This new plugin adds the support for additional file protocols
file:/// included.