Reverted to the Compatibilty mode
Reverted to the Compatibility mode for the Cordova file plugin.
The Internal mode stores the files internally to the app and
makes the private. All other apps cannot access these files
anymore so the WebIntent plugin is useless. Using the inAppBrowser
is limiting because only some of the file types are supported.
Pdf files are not for instance.