BUGFIX: broken links; Resources type html
The Resources type html had broken links.
This commit fixes the issue and introduces a new mechanism.
All the html resources are opened in a inappbrowser window
and the content is filtered in order to ensure the best user
experience.
So if the resource is embedded inside an iframe then the iframe
src attribute is extracted and the inappbrowser window is
pointed to it.
When the resource is not inside a iframe then the page is filtered
and all the unnecessary blocks are stripped.

If there is a redirecting anchor to be clicked it is processed and
again the inappbrowser window is pointed to it.