Contact participants networks: phone, sms & skype
Initial android support for contacting participants through
phone, sms and skype networks. e-Mail network is still available.
Picking the network is prompted when e-Mail is not the only choice.
The common case is when a single participant is selected although
it works with more than one participant provided the selected ones
have more than one network in common.

Renamed the Main controller to Updater which is more appropriate
because of what it actually does.