RecentActivities Moodle 2.7 support + bugfixes
RecentActivity Moodle 2.7 support. There is some retrocompatibilty
code still in place though.

CalendarEvents support in the RecentActivity list.
Starting from Moodle 2.7 the calendar events are now supported.
Clicking on the RecentActivity calendar event entry is going to
show the Calendar view.

Calendar view: the description is now shown when the user
clicks on the event. The event name is always shown.

Bugfixes in the Update mecanism. The discussions were not removed
properly.