Implemented the content update monitor
When the network is not available the app will periodically
check if the network status has changed. If this is the case
then the app will attempt the authentication.

Bugfix: The course navigator was not shown when the network
was off-line. This commit fixes the issue and adds a new feature.