Client side AAI authentication; Clearing cookies
The AAI authentication now works client side. This has been
implemented for security reasons. The AAI account has been meant to
avoid sending the username and password to the service providers such
as Moodle. The course tokens are now requested as the step following
the authentication process.

A new cordova plugin has been added: Phonegap-Cookies-Plugin.
This new plugin clears the cookies set in the browser object allowing
the user to terminate the session and start a new one.