Bugfix: Creating Forum Discussions and Posts
Creating Forum Discussions and Posts blocked the app.
This commit fixes the issue and improves the updating of the forum
related contents. The content is updated everytime a forum and
a discussion is selected.

Another issue was due to the Course sync. Forum contents were
wiped out after sync of a different course. This commit addresses
also that issue.