Scorm fix: The Standard scorm support was broken
The Standard scorm was broken due to the changes in the how the
file paths are managed by the cordova file plugin.
The internal scorm pages loading was also broken because the file
paths passed internally were not complete and absolute.

Also the location.href alteration checks were commented previously.
This commit fixes the issues above.